horizontally too

+++ muxec.py
@@ -12,7 +12,7 @@ import select
 
 STATUS_HEIGHT = 1
 stdScr = curses.initscr()
-scr_height, scr_width = stdScr.getmaxyx()
+full_height, full_width = stdScr.getmaxyx()
 panes = []
 total = 0
 completed_processes = set()
@@ -20,43 +20,58 @@ failed_processes = set()
 all_processes = []
 exhausted = False
 log_file = "/tmp/muxec.log"
+should_log = os.environ.get("MUXEC_LOG") is not None
 
-if os.path.exists(log_file):
+if should_log and os.path.exists(log_file):
     os.remove(log_file)
 
 
 def _log(message):
+    if not should_log:
+        return
     with open(log_file, "a") as f:
         f.write(message + "\n")
 
 
 def write_row(top_offset):
-    _log(f"Adding row at {top_offset}")
-    stdScr.addstr(top_offset, 0, '-' * scr_width)
+    _log(f"Writing full row at {top_offset}")
+    stdScr.addstr(top_offset, 0, '-' * full_width)
 
 
-def create_pane(width, top_offset, bottom_offset):
-    left_offset = 0
-    height = bottom_offset - top_offset + 1
+def write_column(left_offset):
+    _log(f"Writing full column at {left_offset}")
+    for line in range(full_height):
+        if line > 0:
+            stdScr.addstr(line, left_offset, '|')
+
+
+def create_pane(width, height, top_offset, left_offset):
     pad = curses.newpad(height, width)
     pad.scrollok(True)
-    coords = [top_offset, left_offset, bottom_offset, width]
-    _log(f"Adding pane with coords: {coords}")
+    bottom_offset = top_offset + height
+    coords = [top_offset, left_offset, bottom_offset, left_offset + width]
+    _log(f"Adding pane with coords: {coords} (h={height},w={width})")
     return {
         'pad': pad,
         'height': height,
+        'width': width,
         'coords': coords
     }
+
+
+def refresh_pane(pane, y, x):
+    view_top = max(y - pane['height'], 0)
+    view_left = max(x - pane['width'], 0)
+    pad = pane['pad']
+    pad.refresh(view_top, view_left, *pane['coords'])
 
 
 def write_to_pane(pane_num, text):
     pane = panes[pane_num]
     pad = pane['pad']
     y, x = pad.getyx()
-    _log(f"Writing '{text[0:-1]}' at {y},{x}")
     pad.addstr(y, x, text)
-    view_top = max(y - pane['height'], 0)
-    pad.refresh(view_top, 0, *pane['coords'])
+    refresh_pane(pane, y, x)
 
 
 def clear_pane(pane_num):
@@ -67,25 +82,42 @@ def clear_pane(pane_num):
     for line in range(y):
         pad.addstr(line, 0, " " * x)
     pad.move(0, 0)
-    view_top = max(y - pane['height'], 0)
-    pad.refresh(view_top, 0, *pane['coords'])
+    refresh_pane(pane, y, x)
 
 
 def build_views(num_panes):
     curses.noecho()
     curses.cbreak()
     curses.curs_set(0)
-
-    pane_height = (scr_height - STATUS_HEIGHT) // num_panes - 1
-
     update_status()
 
-    for i in range(num_panes):
-        top_offset = STATUS_HEIGHT + (pane_height + 1) * i
-        write_row(top_offset)
-        bottom_offset = top_offset + pane_height
-        top_offset += 1
-        panes.append(create_pane(scr_width, top_offset, bottom_offset))
+    n_rows = 1
+    n_cols = num_panes
+    if num_panes % 3 == 0:
+        n_rows = 3
+        n_cols = num_panes // 3
+    elif num_panes % 2 == 0:
+        n_rows = 2
+        n_cols = num_panes // 2
+
+    pane_height = (full_height - STATUS_HEIGHT) // n_rows
+    pane_width = full_width // n_cols
+
+    _log(f"Setting up grid of {n_rows} rows x {n_cols} cols, pane h={pane_height},w={pane_width}")
+
+    for col in range(n_cols):
+        if col > 0:
+            write_column(col * pane_width)
+    write_row(1)
+    for row in range(n_rows):
+        if row > 0:
+            write_row(STATUS_HEIGHT + row * pane_height)
+        for col in range(n_cols):
+            top_offset = STATUS_HEIGHT + pane_height * row + 1
+            left_offset = pane_width * col
+            if col > 0:
+                left_offset += 1
+            panes.append(create_pane(pane_width - 1, pane_height - 1, top_offset, left_offset))
 
     stdScr.refresh()
 
@@ -94,7 +126,7 @@ def update_status():
     status = f"Running... {len(completed_processes)} / {total} completed"
     if len(failed_processes) > 0:
         status = f"{status}, {len(failed_processes)} failed"
-    stdScr.addstr(0, 0, status + " " * (scr_width - len(status)))
+    stdScr.addstr(0, 0, status + " " * (full_width - len(status)))
     stdScr.refresh()
 
 
@@ -128,22 +160,22 @@ def run(commands):
     active_fds = set()
     data_template = {}
 
-    def _try_run_process(pane_num, clear):
+    def _try_run_process(run_on_pane_num, clear):
         global exhausted
         process = next(gen, None)
         if process is None:
             exhausted = True
             return
         fds = process_fds(process)
-        _log(f"started process '{process.args}' ({process.pid}) with fds {fds} on pane {pane_num}")
+        _log(f"Started process '{process.args}' ({process.pid}) with fds {fds} on pane {run_on_pane_num} (clear={clear})")
         all_processes.append(process)
         for fd in fds:
             active_fds.add(fd)
-            pane_num_by_fd[fd] = pane_num
+            pane_num_by_fd[fd] = run_on_pane_num
             data_template[fd] = bytes()
             process_by_fd[fd] = process
         if clear:
-            clear_pane(pane_num)
+            clear_pane(run_on_pane_num)
 
     def _on_fd_inactive(fd):
         active_fds.remove(fd)
@@ -154,11 +186,12 @@ def run(commands):
         completed_processes.add(process.pid)
         process.poll()
         exit_code = process.returncode
-        _log(f"Marking {pane_num} as free since process {process.pid} completed with code {exit_code}")
+        ready_pane_num = pane_num_by_fd[fd]
+        _log(f"Process {process.pid} at {ready_pane_num} completed with code {exit_code}")
         if exit_code != 0:
             failed_processes.add(process.pid)
         update_status()
-        return pane_num
+        return ready_pane_num
 
     for pane_num in range(len(panes)):
         _try_run_process(pane_num, False)
@@ -218,17 +251,17 @@ def parse_args():
 def main():
     opts = parse_args()
     commands = opts.commands
-    parallelism = opts.parallelism
-    _log(f"running {len(commands)} commands with {parallelism} parallelism, terminal is h={scr_height}, w={scr_width}")
+    parallelism = min(opts.parallelism, len(commands))
+    _log(f"running {len(commands)} commands with {parallelism} parallelism, terminal is h={full_height}, w={full_width}")
 
     num_panes = parallelism
 
     global total
     total = len(commands)
-    build_views(num_panes)
 
     failed = False
     try:
+        build_views(num_panes)
         run(commands)
     except KeyboardInterrupt:
         print("interrupted, shutting down...")
@@ -250,6 +283,8 @@ def main():
         print(f"Completed running {len(all_processes)} processes, {len(failed_processes)} failed")
         for process in all_processes:
             print(f"\tProcess '{process.args}' ({process.pid}) completed with code {process.returncode}")
+    else:
+        print("internal error")
 
 
 if __name__ == '__main__':
